+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - fixed merging of user config with default config. Rules are merged, for all other options the user config is overwriting the default configuration.
+- broken link rule no longer fails for a tag without href
 
 ## 1.0.0-rc6